requirements.txt
flask
flask-mysqldb
flask-wtf
pyyaml
flask-mail
nexmo